+++ io_scene_niftools/modules/nif_export/property/texture/types/bsshadertexture.py
@@ -39,6 +39,7 @@
 from pyffi.formats.nif import NifFormat
 
 from io_scene_niftools.modules.nif_export.property.texture import TextureWriter, TextureSlotManager
+from io_scene_niftools.utils.consts import TEX_SLOTS
 
 
 class BSShaderTexture(TextureSlotManager):
@@ -63,13 +64,13 @@ class BSShaderTexture(TextureSlotManager):
     def export_bs_effect_shader_prop_textures(self, bsshader):
         bsshader.texture_set = self._create_textureset()
 
-        if self.slots["Base"]:
-            bsshader.source_texture = TextureWriter.export_texture_filename(self.slots["Base"])
-        if self.slots["Glow"]:
-            bsshader.greyscale_texture = TextureWriter.export_texture_filename(self.slots["Glow"])
+        if self.slots[TEX_SLOTS.BASE]:
+            bsshader.source_texture = TextureWriter.export_texture_filename(self.slots[TEX_SLOTS.BASE])
+        if self.slots[TEX_SLOTS.GLOW]:
+            bsshader.greyscale_texture = TextureWriter.export_texture_filename(self.slots[TEX_SLOTS.GLOW])
 
-        # clamp Mode
-        bsshader.texture_clamp_mode = 65283
+        # get the offset, scale and UV wrapping mode and set them
+        self.export_uv_transform(bsshader)
 
     def export_bs_lighting_shader_prop_textures(self, bsshader):
         texset = self._create_textureset()
@@ -79,34 +80,14 @@ class BSShaderTexture(TextureSlotManager):
         texset.num_textures = 9
         texset.textures.update_size()
 
-        if self.slots["Detail"]:
-            texset.textures[6] = TextureWriter.export_texture_filename(self.slots["Detail"])
+        if self.slots[TEX_SLOTS.DETAIL]:
+            texset.textures[6] = TextureWriter.export_texture_filename(self.slots[TEX_SLOTS.DETAIL])
 
-        if self.slots["Gloss"]:
-            texset.textures[7] = TextureWriter.export_texture_filename(self.slots["Gloss"])
+        if self.slots[TEX_SLOTS.GLOSS]:
+            texset.textures[7] = TextureWriter.export_texture_filename(self.slots[TEX_SLOTS.GLOSS])
 
-        # TODO [shader] UV offset node support
-        # # UV Offset
-        # if hasattr(bsshader, 'uv_offset'):
-        #     self.export_uv_offset(bsshader)
-        #
-        # # UV Scale
-        # if hasattr(bsshader, 'uv_scale'):
-        #     self.export_uv_scale(bsshader)
-
-        # Texture Clamping mode
-        b_img = self.slots["Base"].image
-        # TODO [texture] Implement clamp on image wrapping
-        # if not b_img.use_clamp_x:
-        #     wrap_s = 2
-        # else:
-        #     wrap_s = 0
-        # if not b_img.use_clamp_y:
-        #     wrap_t = 1
-        # else:
-        #     wrap_t = 0
-        #
-        # bsshader.texture_clamp_mode = (wrap_s + wrap_t)
+        # get the offset, scale and UV wrapping mode and set them
+        self.export_uv_transform(bsshader)
 
     def export_bs_shader_pp_lighting_prop_textures(self, bsshader):
         bsshader.texture_set = self._create_textureset()
@@ -114,28 +95,63 @@ class BSShaderTexture(TextureSlotManager):
     def _create_textureset(self):
         texset = NifFormat.BSShaderTextureSet()
 
-        if self.slots["Base"]:
-            texset.textures[0] = TextureWriter.export_texture_filename(self.slots["Base"])
+        if self.slots[TEX_SLOTS.BASE]:
+            texset.textures[0] = TextureWriter.export_texture_filename(self.slots[TEX_SLOTS.BASE])
 
-        if self.slots["Normal"]:
-            texset.textures[1] = TextureWriter.export_texture_filename(self.slots["Normal"])
+        if self.slots[TEX_SLOTS.NORMAL]:
+            texset.textures[1] = TextureWriter.export_texture_filename(self.slots[TEX_SLOTS.NORMAL])
 
-        if self.slots["Glow"]:
-            texset.textures[2] = TextureWriter.export_texture_filename(self.slots["Glow"])
+        if self.slots[TEX_SLOTS.GLOW]:
+            texset.textures[2] = TextureWriter.export_texture_filename(self.slots[TEX_SLOTS.GLOW])
 
-        if self.slots["Detail"]:
-            texset.textures[3] = TextureWriter.export_texture_filename(self.slots["Detail"])
+        if self.slots[TEX_SLOTS.DETAIL]:
+            texset.textures[3] = TextureWriter.export_texture_filename(self.slots[TEX_SLOTS.DETAIL])
 
         return texset
 
-    def export_uv_offset(self, shader):
-        shader.uv_offset.u = self.slots["Base"].offset.x
-        shader.uv_offset.v = self.slots["Base"].offset.y
+    def export_uv_transform(self, shader):
+        # get the offset, scale and UV wrapping mode and set them
+        x_scale, y_scale, x_offset, y_offset, clamp_x, clamp_y = self.get_global_uv_transform_clip()
+        # default values for if they haven't been defined:
+        if x_scale is None:
+            x_scale = 1
+        if y_scale is None:
+            y_scale = 1
+        if x_offset is None:
+            x_offset = 0
+        if y_offset is None:
+            y_offset = 0
+        else:
+            # need to translate blender offset to nif offset to get the same results
+            y_offset = 1 - y_scale - y_offset
+        if clamp_x is None:
+            clamp_x = False
+        if clamp_y is None:
+            clamp_y = False
 
-        return shader
+        if hasattr(shader, "uv_scale"):
+            shader.uv_scale.u = x_scale
+            shader.uv_scale.v = y_scale
 
-    def export_uv_scale(self, shader):
-        shader.uv_scale.u = self.slots["Base"].scale.x
-        shader.uv_scale.v = self.slots["Base"].scale.y
+        if hasattr(shader, 'uv_offset'):
+            shader.uv_offset.u = x_offset
+            shader.uv_offset.v = y_offset
+
+        # Texture Clamping mode
+        if hasattr(shader, 'texture_clamp_mode'):
+            if self.slots[TEX_SLOTS.BASE] and (self.slots[TEX_SLOTS.BASE].extension == "CLIP"):
+                # if the extension is clip, we know the wrap mode is clamp for both,
+                shader.texture_clamp_mode = (shader.texture_clamp_mode - shader.texture_clamp_mode % 256) + NifFormat.TexClampMode.CLAMP_S_CLAMP_T
+            else:
+                # otherwise, look at the given clip modes from the nodes
+                if not clamp_x:
+                    wrap_s = 2
+                else:
+                    wrap_s = 0
+                if not clamp_y:
+                    wrap_t = 1
+                else:
+                    wrap_t = 0
+                shader.texture_clamp_mode = (shader.texture_clamp_mode - shader.texture_clamp_mode % 256) + (wrap_s + wrap_t)
 
         return shader